Refactor backend validation handler.
