+++ server/services/alcoholMonitoringService.ts
@@ -3,10 +3,9 @@ import RestClient from '../data/restClient'
 import { AuthenticatedRequestInput } from '../interfaces/request'
 import AlcoholMonitoringModel, { AlcoholMonitoring } from '../models/AlcoholMonitoring'
 import { AlcoholMonitoringFormData, AlcoholMonitoringFormDataValidator } from '../models/form-data/alcoholMonitoring'
-import { ValidationResult, ValidationResultModel } from '../models/Validation'
+import { ValidationResult } from '../models/Validation'
 import { SanitisedError } from '../sanitisedError'
-import { convertZodErrorToValidationError } from '../utils/errors'
-import processBackendValidationErrors from '../utils/validators/processBackendValidationErrors'
+import { convertZodErrorToValidationError, convertBackendErrorToValidationError } from '../utils/errors'
 
 type AlcoholMonitoringInput = AuthenticatedRequestInput & {
   orderId: string
@@ -29,11 +28,10 @@ export default class AlcoholMonitoringService {
       if (e instanceof ZodError) {
         return convertZodErrorToValidationError(e)
       }
-      const sanitisedError = e as SanitisedError
 
+      const sanitisedError = e as SanitisedError
       if (sanitisedError.status === 400) {
-        const parsedErrors = ValidationResultModel.parse((e as SanitisedError).data)
-        return processBackendValidationErrors(parsedErrors)
+        return convertBackendErrorToValidationError(sanitisedError)
       }
 
       throw e
