+++ server/services/monitoringConditionsService.ts
@@ -2,14 +2,13 @@ import { ZodError } from 'zod'
 import RestClient from '../data/restClient'
 import { AuthenticatedRequestInput } from '../interfaces/request'
 import MonitoringConditionsModel, { MonitoringConditions } from '../models/MonitoringConditions'
-import { ValidationResult, ValidationResultModel } from '../models/Validation'
+import { ValidationResult } from '../models/Validation'
 import { SanitisedError } from '../sanitisedError'
-import { convertZodErrorToValidationError } from '../utils/errors'
+import { convertZodErrorToValidationError, convertBackendErrorToValidationError } from '../utils/errors'
 import {
   MonitoringConditionsFormData,
   validateMonitoringConditionsFormData,
 } from '../models/form-data/monitoringConditions'
-import processBackendValidationErrors from '../utils/validators/processBackendValidationErrors'
 
 export type UpdateMonitoringConditionsInput = AuthenticatedRequestInput & {
   orderId: string
@@ -36,8 +35,7 @@ export default class MonitoringConditionsService {
 
       const sanitisedError = e as SanitisedError
       if (sanitisedError.status === 400) {
-        const parsedErrors = ValidationResultModel.parse((e as SanitisedError).data)
-        return processBackendValidationErrors(parsedErrors)
+        return convertBackendErrorToValidationError(sanitisedError)
       }
 
       throw e
