+++ server/services/trailMonitoringService.ts
@@ -3,10 +3,9 @@ import RestClient from '../data/restClient'
 import { AuthenticatedRequestInput } from '../interfaces/request'
 import { TrailMonitoringFormData, TrailMonitoringFormDataValidator } from '../models/form-data/trailMonitoring'
 import TrailMonitoringModel, { TrailMonitoring } from '../models/TrailMonitoring'
-import { ValidationResult, ValidationResultModel } from '../models/Validation'
+import { ValidationResult } from '../models/Validation'
 import { SanitisedError } from '../sanitisedError'
-import { convertZodErrorToValidationError } from '../utils/errors'
-import processBackendValidationErrors from '../utils/validators/processBackendValidationErrors'
+import { convertZodErrorToValidationError, convertBackendErrorToValidationError } from '../utils/errors'
 
 type TrailMonitoringInput = AuthenticatedRequestInput & {
   orderId: string
@@ -29,11 +28,12 @@ export default class TrailMonitoringService {
       if (e instanceof ZodError) {
         return convertZodErrorToValidationError(e)
       }
+
       const sanitisedError = e as SanitisedError
       if (sanitisedError.status === 400) {
-        const parsedErrors = ValidationResultModel.parse((e as SanitisedError).data)
-        return processBackendValidationErrors(parsedErrors)
+        return convertBackendErrorToValidationError(sanitisedError)
       }
+
       throw e
     }
   }
